refactor: menyederhanakan proses importing dan kompleksitas kode
Mengapa perubahan dilakukan:
* Menyederhanakan setiap class pada folder Disambiguator menjadi fungsi
dan menyatukannya menjadi satu class mempermudah proses importing dan
maintenance kode.
* Class PrefixDisambiguator dan AbstractPrefixDisambiguator disatukan
karena terlihat redundant dan kompleks.
* Menghapus beberapa method pada VisitorProvider karena kurang Pythonic
